proctoring-service/requirements.txt
# Core Framework (Essential)
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Computer Vision (Phase 2-3 integration)
# Note: Keep heavy models optional; OpenCV is required for current Phase-1 detectors.
opencv-python-headless>=4.10.0
# mediapipe>=0.10.30  
# deepface>=0.0.75

# Numerical Computing
numpy>=2.0.0

# Web utilities
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
python-json-logger>=2.0.0
pillow>=10.2.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx==0.25.1

# Optional: Development
black>=24.0.0
flake8>=7.0.0
mypy>=1.7.0

